requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
Brotli==1.0.9
celery==5.2.7
autobahn==22.12.1
Automat==22.10.0
bidict==0.22.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
# daphne==4.0.0
dj-database-url==1.2.0
dj-rest-auth==2.2.6
Django==4.1.5
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-etc==1.4.0
django-filter==22.1
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-phonenumber-field==7.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
ecdsa==0.18.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
mailjet-rest==1.3.4
MarkupSafe==2.1.1
model-bakery==1.10.1
msgpack==1.0.4
mypy-extensions==0.4.3
# mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.0
pathspec==0.10.3
phonenumbers==8.13.4
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
mypy-extensions==0.4.3
oauthlib==3.2.2
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pyotp==2.8.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-watch==4.2.0
python-decouple==3.7
python-jose==3.3.0
pyOpenSSL==23.0.0
python-engineio==4.3.4
python-socketio==5.7.2
python3-openid==3.2.0
pytz==2022.7
# pywin32==305
PyYAML==6.0
rave-python==1.2.16
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
roundrobin==0.0.4
rsa==4.9
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
watchdog==2.2.1
Werkzeug==2.2.2
websocket-client==1.4.2
whitenoise==6.3.0
zope.interface==5.5.2
